Move the calendar toolbar
