+++ app/views/components/calendar/example-calendar-toolbar.html
