+++ src/core/_controls-new.scss
@@ -123,8 +123,8 @@
 
 // Complex ====/
 @import '../components/calendar/calendar';
-@import '../components/calendar-toolbar/calendar-toolbar';
-@import '../components/calendar-toolbar/calendar-toolbar-new';
+@import '../components/calendar/calendar-toolbar';
+@import '../components/calendar/calendar-toolbar-new';
 @import '../components/calendar/calendar-new';
 @import '../components/compositeform/compositeform';
 @import '../components/contextualactionpanel/contextualactionpanel';
