+++ src/core/_controls.scss
@@ -73,7 +73,7 @@
 
 // Complex ====/
 @import '../components/calendar/calendar';
-@import '../components/calendar-toolbar/calendar-toolbar';
+@import '../components/calendar/calendar-toolbar';
 @import '../components/compositeform/compositeform';
 @import '../components/contextualactionpanel/contextualactionpanel';
 @import '../components/datagrid/datagrid';
